Move package around, reduce binary size
This commit moves package files to a common place so that CLI code will
not import package code inside controller
