+++ cli/cmd/config/cat.go
@@ -6,9 +6,8 @@ import (
 	"fmt"
 	"os"
 
-	"gopkg.in/yaml.v2"
-
 	"github.com/rancher/rio/cli/pkg/clicontext"
+	"gopkg.in/yaml.v2"
 	corev1 "k8s.io/api/core/v1"
 )
