+++ cli/main.go
@@ -6,10 +6,9 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/rancher/rio/cli/cmd/buildhistory"
-
 	"github.com/docker/docker/pkg/reexec"
 	"github.com/rancher/rio/cli/cmd/attach"
+	"github.com/rancher/rio/cli/cmd/buildhistory"
 	"github.com/rancher/rio/cli/cmd/builds"
 	"github.com/rancher/rio/cli/cmd/config"
 	"github.com/rancher/rio/cli/cmd/dashboard"
