+++ cli/pkg/tables/service.go
@@ -5,16 +5,15 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/rancher/rio/cli/cmd/util"
-
 	webhookv1 "github.com/rancher/gitwatcher/pkg/apis/gitwatcher.cattle.io/v1"
+	"github.com/rancher/rio/cli/cmd/util"
 	"github.com/rancher/rio/cli/pkg/table"
-	"github.com/rancher/rio/modules/service/controllers/service/populate/serviceports"
 	riov1 "github.com/rancher/rio/pkg/apis/rio.cattle.io/v1"
 	v1 "github.com/rancher/rio/pkg/apis/rio.cattle.io/v1"
 	"github.com/rancher/rio/pkg/constants"
 	"github.com/rancher/rio/pkg/controllers/pkg"
 	"github.com/rancher/rio/pkg/riofile/stringers"
+	"github.com/rancher/rio/pkg/serviceports"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 )
