+++ modules/service/controllers/service/populate/pod/containers.go
@@ -6,9 +6,9 @@ import (
 	"strings"
 
 	"github.com/docker/distribution/reference"
-	"github.com/rancher/rio/modules/service/controllers/service/populate/serviceports"
 	riov1 "github.com/rancher/rio/pkg/apis/rio.cattle.io/v1"
 	"github.com/rancher/rio/pkg/riofile/stringers"
+	"github.com/rancher/rio/pkg/serviceports"
 	"github.com/rancher/rio/pkg/services"
 	"github.com/rancher/wrangler/pkg/name"
 	v1 "k8s.io/api/core/v1"
