+++ pkg/constants/constants.go
@@ -16,6 +16,13 @@ const (
 	DefaultGitCrendentialSSH = "gitcredential-ssh"
 	DefaultDockerCrendential = "dockerconfig"
 	DefaultGithubCrendential = "githubtoken"
+
+	StackLabel     = "gitwatcher.rio.cattle.io/stack"
+	ServiceLabel   = "gitwatcher.rio.cattle.io/service"
+	ContainerLabel = "gitwatcher.rio.cattle.io/container"
+
+	GitCommitLabel = "gitwatcher.rio.cattle.io/git-commit"
+	LogTokenLabel  = "gitwatcher.rio.cattle.io/log-token"
 )
 
 var (
